# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Twitter Channnel Service",
        "contact": {
            "email": "ojilesanmi7@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/chatbot/subscribe": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Subscribe users to a chat bot",
                "description": "Subscribe users to a chat bot using the specified parameters.",
                "operationId": "7d45e66b2c73a7b1f6ce4fc3d79926c7",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "The user ID.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates whether the user was successfully subscribed to the chatbot",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Invalid User Id.",
                                            "type": "string",
                                            "example": "The User id must be one of the four specified in the document"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/chat/subscribe": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Subscribe users to a chat or a channel",
                "description": "Subscribe users to a chat or a channel using the specified parameters.",
                "operationId": "ed4b8d037c5741553c06f359e352a75f",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "The user ID.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "channelId": {
                                        "description": "The ID of the channel to subscribe the user to.",
                                        "type": "string",
                                        "example": "channel123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates whether the user was successfully subscribed to the chat or channel",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Invalid User Id.",
                                            "type": "string",
                                            "example": "The User id must be one of the four specified in the document"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/bot-user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the account details associated with the app",
                "operationId": "6e9ec2642bed3dda2253696cf27b6a37",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "The user ID.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "80000000-8000-8000-8000-000000000008"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get the account details associated with the app",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates the operation was successful",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "result": {
                                            "description": "The return object from the channel",
                                            "type": "object",
                                            "example": {
                                                "id": "13696209441431539484",
                                                "name": "Opeyemi Ilesanmi",
                                                "username": "ilejohn"
                                            }
                                        },
                                        "message": {
                                            "description": "A message indicating success",
                                            "type": "string",
                                            "example": "User retrieved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/me": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Get the user details of the auth user",
                "operationId": "f9d3500239c7e5b651107751e95acab6",
                "responses": {
                    "200": {
                        "description": "Get the user details of the auth user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates the operation was successful",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "result": {
                                            "description": "The user details",
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Opeyemi Ilesanmi",
                                                "twitter_id": "493373737323343",
                                                "twitter_screen_name": "marvis",
                                                "twitter_oauth_token": "xxxxxxxxxxxxxxxxxxxxxxxxx",
                                                "twitter_oauth_token_secret": "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
                                                "email": null,
                                                "email_verified_at": null,
                                                "created_at": "2020-10-10T16:43:15.000000Z",
                                                "updated_at": "2020-10-10T16:43:15.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The error why it wasn't successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "An error message why the action was not successful",
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tweet": {
            "post": {
                "tags": [
                    "Communication"
                ],
                "summary": "Post a tweet from the account associated with the app",
                "description": "Post a tweet from the account associated with the app using the specified parameters.",
                "operationId": "f7f21f4238ff193f943c47f4a186902b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "description": "The text to be tweeted",
                                        "type": "string",
                                        "example": "A random tweet from a laravel app"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates whether the action was successful",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "result": {
                                            "description": "The return object",
                                            "type": "object",
                                            "example": {
                                                "edit_history_tweet_ids": "1727801020732801098",
                                                "id": "1727801020732801098",
                                                "text": "A random tweet from a laravel app"
                                            }
                                        },
                                        "message": {
                                            "description": "A message indicating success",
                                            "type": "string",
                                            "example": "Tweet successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "An error message why the action was not successful",
                                            "type": "string",
                                            "example": "Error message"
                                        },
                                        "success": {
                                            "description": "Indicates the operation was unsuccessful",
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}